Work around timing issue in TestHelpers' check_storage_framework_files()
The issue is that we don't reliably know which created backup tarfile
corresponds to which backup task. The short-term workaround is to do
a brute-force NxM comparison until we find a match.

Once we get Restore working, a better way to do this test is to backup +
restore and then compare the source with the restore.